Fixed attention object restoring when reentering whisper mode
It wasn't working as I thought it was and could potentially crash in scenarios like Framing Frame 2 where bags spawn at the beginning of the heist, when going loud.

Also nuking the set_rescue_state function since rescue_state() always returns true, which means assaults are not supposed to disable it for current hostages. This can cause other issues as well.