Bot bag linking voiceline is only used on throws
Again.

I forgot that when a bot gets downed while carrying a bag, and then gets revived, the bag is linked once again. And due to how I changed this recently, it would cause the bot to use this voiceline the instant they get revived, along with their revive thanking one.