Bullet Hell now loads 6 bullets into the magazine from your total ammo when triggered.
The skill felt strong on LMGs but undertuned on SMGs due to how finnicky it can be to proc- and often when you did trigger it you already wasted most of your magazine.
	This change should minimally effect its synergy with LMGs while making it feel more consistent on SMGs.
	It also reinforces the fantasy of just being an absolute bullet hose when you point your gun down a hallway with a bunch of enemies nicely lined up.
	NOTE: This moves ammo you already have into your magazine. It does not restore ammo.
	Also note: This occurs *ever* time you kill multiple enemies in full auto without releasing the trigger.

Might be overtuned, REALLY needs playtesting + feedback. If it is, I'll start by slightly reducing the other buffs first since if it's a bit much on SMGs then it's definitely a bit much on LMGs.

Also worth double checking that its synergy with auto-shotguns isn't absolutely busted-strong, since they tend to have far smaller magsizes and spread makes proccing the effect easier.