Bullet Hell Buff, 90 damage rifles get -5 stability across the board.
The previous nerf to bullet hell was the result of it having an incorrect free ammo chance, since that has been fixed it can be bumped up to be where T4 damage skills should be.

Was playing with 90 damage rifles and they really felt overtuned relative to their lower damage peers.
	Harsher penalties for low handling + a minor bump down on their own handling should help check their power.
	60 damage rifles are probably fine, since they aren't as big of a jump from 48 damage and have a marginally awkward ammo economy.