Removed sound environment changes to save ear drums.
The file we had for this was a direct copy/paste of the Payday 009 luadump.

The only change this made in practice over vanilla is that self._ambience_sources_count was set to 4 in our code. The rest was effectively duplicate (but slightly less polished) vanilla code.
What this meant in practice is that global ambient sounds became 4x as loud.

While it would be nice having better map audio, it's better suited to a standalone mod. And regardless, this approach of just yeeting in the entire old file is not a good way to go.