Rodion 3B
Could have used volley fire, but when looking into the gun it appeared that it was designed to always fire each barrel equally.
So instead it's a 48 damage chonky AR that fires 3 rays at once in both auto and single fire, giving it an "RPM" of 1800 in matching with most specs I can find.
Since the barrels are arranged horizontally, thought it would be cute to make it have more horizontal than vertical spread.
	Also covers the bulk of the code needed for the leveller attachment.